server/requirements.txt
flask>=3.0
requests>=2.31
chromadb>=0.5
Pillow>=10.0
